refactor: migrate Event type to zod schema and remove legacy event type
