--- types/event.ts
@@ -1,8 +1,0 @@
-export interface Event {
-  id: string;
-  title: string;
-  startDate: string;
-  endDate: string;
-  description?: string | null;
-  status?: "pending" | "completed" | "cancelled";
-}
